disable cloud discovery by default
SiliconDust staff have recommended that one
does not use the cloud based discovery API
(by default) due to various issues with ISPs
(typically CGNAT), or the various latencies
for changes (adds/deletes).  Instead the
recommendation is that one should typically
use (local) broadcast discovery (and/or SSDP)
with a fallback option of using the cloud
discovery if needed due to local networking
requirements.  This patch disables cloud
discovery by default and adds in an option
(--use-cloud-discovery) to enable it.